comments soft delete done

+++ src/blogpages/crudComments.php
@@ -4,7 +4,7 @@ session_start();
 $userId = $_SESSION['userId'];
 $data = json_decode(file_get_contents("php://input"), true);
 
-//********************** Add tag **************************// 
+//********************** Add comment **************************// 
 
 if (!empty($data['commentContent'])) {
     $commentContent = htmlspecialchars($data['commentContent']);
@@ -19,13 +19,27 @@ if (!empty($data['commentContent'])) {
     echo "Comment sent successfully!";
     exit;
 }
+//********************** Update comment **************************//
 
-//********************** Delete tag **************************// 
+if (!empty($data['newComment'])) {
+    $commentContent = htmlspecialchars($data['newComment']);
+    $commentId = $data['commentId2'];
+
+    $update = "UPDATE comments SET commentContent = ? WHERE commentId = ?";
+    $stmt = $conn->prepare($update);
+    $stmt->bind_param("si", $commentContent, $commentId);
+    $stmt->execute();
+    $stmt->close();
+    echo "Comment updated successfully!";
+    exit;
+}
+
+//********************** Delete comment **************************// 
 
 if (!empty($data['commentId'])) {
     $commentId =  $data['commentId'];
 
-    $delete = "DELETE FROM comments WHERE commentId = ?";
+    $delete = "UPDATE comments SET isDeleted = 1 WHERE commentId = ?";
     $stmt = $conn->prepare($delete);
     $stmt->bind_param("i", $commentId);
     $stmt->execute();
@@ -34,31 +48,21 @@ if (!empty($data['commentId'])) {
     exit;
 }
 
-//********************** Update tag **************************//
+if (!empty($data['commentId2'])) {
+    $commentId =  $data['commentId2'];
 
-// if (!empty($data['tagName2'])) {
-//     $tagId = $data['tagId3'];
-//     $tagName = $data['tagName2'];
+    $delete = "UPDATE comments SET isDeleted = 0 WHERE commentId = ?";
+    $stmt = $conn->prepare($delete);
+    $stmt->bind_param("i", $commentId);
+    $stmt->execute();
+    $stmt->close();
+    echo "Comment reverted successfully!";
+    exit;
+}
 
-//     $checkQuery = "SELECT COUNT(*) as count FROM tags WHERE tagName = ?";
-//     $stmt = $conn->prepare($checkQuery);
-//     $stmt->bind_param("s", $tagName);
-//     $stmt->execute();
-//     $result2 = $stmt->get_result();
-//     $tagCount2 = $result2->fetch_assoc()['count'];
+//********************** View comments **************************//
 
-//     if ($tagCount2 > 0) {
-//         return false;
-//     } else {
-//         $update = "UPDATE tags SET tagName = ? WHERE tagId = ?";
-//         $stmt = $conn->prepare($update);
-//         $stmt->bind_param("si", $tagName, $tagId);
-//         $stmt->execute();
-//         $stmt->close();
-//         echo "Tag inserted successfully!";
-//     }
-//     exit;
-// }
+
 if (!empty($data['articleId2'])) {
     $articleId = $data['articleId2'];
     $select = "SELECT * FROM comments WHERE articleId = $articleId";
@@ -69,31 +73,53 @@ if (!empty($data['articleId2'])) {
             $commentId = $row['commentId'];
             $userSession = $row['userSession'];
             $commentContent = htmlspecialchars_decode($row['commentContent']);
+            $isDeleted = $row['isDeleted'];
             $user = "SELECT * FROM users WHERE userId = $userSession";
             $result2 = mysqli_query($conn, $user);
             $row2 = mysqli_fetch_assoc($result2);
             $userName = $row2['userName'];
-
-            echo '
-                <div class="flex flex-col w-full shadow-lg border-t-2 p-2 pl-4">
-                    <div class="flex w-full justify-between">
-                        <h1 class="text-gray-500"><i class="bx bx-user text-gray-500 text-xl border-gray-500"></i>' . $userName . '</h1>
-                        ';
-            if ($userId == $userSession) {
-                echo '<div>
-                            <i onclick="openpopup(' . $commentId . ',' . $articleId . ');" class="bx bx-edit-alt text-gray-500 text-xl border-gray-500 cursor-pointer"></i>
-                            <i onclick="deleteComment(' . $commentId . ')" class="bx bx-message-alt-x text-gray-500 text-xl border-gray-500 cursor-pointer"></i>
-                          </div>';
-            } else if (isset($_SESSION['admin_name']) || isset($_SESSION['administrator_name'])) {
+            if (($isDeleted == 1 && $userId == $userSession) || isset($_SESSION['admin_name']) || isset($_SESSION['administrator_name'])) {
                 echo '
+                <div id="comment' . $commentId . '" class="flex flex-col w-full shadow-lg border-t-2 p-2 pl-4 bg-red-500/30">
+                    <div class="flex w-full justify-between">
+                        <h1 id="user' . $commentId . '" class="text-gray-500"><i class="bx bx-user text-gray-500 text-xl border-gray-500"></i>' . $userName . '</h1>
+                        ';
+                if ($userId == $userSession) {
+                    echo '
                             <div>
-                                <i onclick="openpopup(' . $commentId . ',' . $articleId . ');" class="bx bx-edit-alt text-gray-500 text-xl border-gray-500 cursor-pointer"></i>
+                                <p onclick="undoDelete(' . $commentId . ',' . $articleId . ')" class="cursor-pointer underline text-gray-500">Undo</p>
+                            </div>';
+                } else if (isset($_SESSION['admin_name']) || isset($_SESSION['administrator_name'])) {
+                    echo '
+                            <div>
+                                <p onclick="undoDelete(' . $commentId . ',' . $articleId . ')" class="cursor-pointer underline text-gray-500">Undo</p>
+                            </div>';
+                }
+                echo '</div>
+                    <p id="p' . $commentId . '">[Deleted comment]</p>
+                </div>';
+            } else {
+                echo '
+                <div id="comment' . $commentId . '" class="flex flex-col w-full shadow-lg border-t-2 p-2 pl-4">
+                    <div class="flex w-full justify-between">
+                        <h1 id="user' . $commentId . '" class="text-gray-500"><i class="bx bx-user text-gray-500 text-xl border-gray-500"></i>' . $userName . '</h1>
+                        ';
+                if ($userId == $userSession) {
+                    echo '<div>
+                            <i onclick="editComment(' . $commentId . ',' . $articleId . ');" class="bx bx-edit-alt text-gray-500 text-xl border-gray-500 cursor-pointer"></i>
+                            <i onclick="deleteComment(' . $commentId . ',' . $articleId . ')" class="bx bx-message-alt-x text-gray-500 text-xl border-gray-500 cursor-pointer"></i>
+                          </div>';
+                } else if (isset($_SESSION['admin_name']) || isset($_SESSION['administrator_name'])) {
+                    echo '
+                            <div>
+                                <i onclick="editComment(' . $commentId . ',' . $articleId . ');" class="bx bx-edit-alt text-gray-500 text-xl border-gray-500 cursor-pointer"></i>
                                 <i onclick="deleteComment(' . $commentId . ',' . $articleId . ')" class="bx bx-message-alt-x text-gray-500 text-xl border-gray-500 cursor-pointer"></i>
                             </div>';
-            }
-            echo '</div>
-                    <p>' . $commentContent . '</p>
+                }
+                echo '</div>
+                    <p id="p' . $commentId . '">' . $commentContent . '</p>
                 </div>';
+            }
         }
     } else {
         echo '<div class="flex flex-col w-full shadow-md rounded-lg border-t-2 p-2 pl-4 text-center">
